+++ labs/raycaster/polygon.h
@@ -2,6 +2,7 @@
 #define POLYGON_H
 
 #include "ray.h"
+#include "utils.h"
 
 #include <QPointF>
 #include <cmath>
@@ -10,8 +11,6 @@
 #include <optional>
 #include <utility>
 #include <vector>
-
-using Intersection = std::optional<std::pair<QPointF, double>>;
 
 class Polygon {
    public:
@@ -58,27 +57,6 @@ class Polygon {
             intersection && (!result || intersection->second < result->second)) {
             result = intersection;
         }
-    }
-
-    static Intersection IntersectRay(
-        const QPointF& vertex1, const QPointF& vertex2, const Ray& ray) {
-        if (Ray::AreParallel(ray, {vertex2, vertex1})) {
-            return std::nullopt;
-        }
-        const auto [r_px, r_py] = ray.GetBegin();
-        const auto r_dx = std::cos(ray.GetAngle());
-        const auto r_dy = std::sin(ray.GetAngle());
-        const auto [s_px, s_py] = vertex1;
-        const auto [s_dx, s_dy] = vertex2 - vertex1;
-        const auto t2 = (r_dx * (s_py - r_py) + r_dy * (r_px - s_px)) / (s_dx * r_dy - s_dy * r_dx);
-        auto t1 = (s_px + s_dx * t2 - r_px) / r_dx;
-        if (std::abs(r_dx) < kEpsilon) [[unlikely]] {
-            t1 = (s_py + s_dy * t2 - r_py) / r_dy;
-        }
-        if (t1 < -kEpsilon || t2 < -kEpsilon || t2 > 1 + kEpsilon) {
-            return std::nullopt;
-        }
-        return std::make_pair(ray.GetBegin() + QPointF(r_dx, r_dy) * t1, t1);
     }
 };
